requirements.txt
fastapi
uvicorn[standard]

# LangChain Core + Ecosystem
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-google-genai

# Vector Store + Embeddings
faiss-cpu
sentence-transformers

# Document Processing
pypdf
python-dotenv

# LLMs / Agents
openai
langgraph